Move resize implementation to separate file
