+++ modules/imgproc/src/imgwarp.hpp
@@ -56,8 +56,6 @@ namespace cv
 namespace opt_AVX2
 {
 #if CV_TRY_AVX2
-void resizeNN2_AVX2(const Range&, const Mat&, Mat&, int*, int, double);
-void resizeNN4_AVX2(const Range&, const Mat&, Mat&, int*, int, double);
 int warpAffineBlockline(int *adelta, int *bdelta, short* xy, short* alpha, int X0, int Y0, int bw);
 #endif
 }
@@ -65,10 +63,6 @@ int warpAffineBlockline(int *adelta, int *bdelta, short* xy, short* alpha, int X
 namespace opt_SSE4_1
 {
 #if CV_TRY_SSE4_1
-void resizeNN2_SSE4_1(const Range&, const Mat&, Mat&, int*, int, double);
-void resizeNN4_SSE4_1(const Range&, const Mat&, Mat&, int*, int, double);
-
-int VResizeLanczos4Vec_32f16u_SSE41(const uchar** _src, uchar* _dst, const uchar* _beta, int width);
 void convertMaps_nninterpolate32f1c16s_SSE41(const float* src1f, const float* src2f, short* dst1, int width);
 void convertMaps_32f1c16s_SSE41(const float* src1f, const float* src2f, short* dst1, ushort* dst2, int width);
 void convertMaps_32f2c16s_SSE41(const float* src1f, short* dst1, ushort* dst2, int width);
